+++ routes/auth.js
@@ -6,7 +6,7 @@ const {
     signin,
     signout,
     requireSignin 
-} = require('../controllers/user')
+} = require('../controllers/auth')
 const { userSignupValidator } = require('../validator')
 
 router.post('/signup', userSignupValidator, signup)
